requirements.txt
mysqlclient==1.4.2.post1
Django==2.2.1
djangorestframework==3.9.4
django-cors-middleware==1.3.1
djangorestframework-jwt==1.11.0
django-filter==2.1.0